requirements.txt
numpy
scipy
pandas
matplotlib
networkx
python-igraph
